rewrite to make sure everything can be tested properly
