+++ pom.xml
@@ -32,7 +32,7 @@
     <description>Manage one or more BagStores</description>
     <inceptionYear>2016</inceptionYear>
     <properties>
-        <main-class>nl.knaw.dans.easy.bagstore.Command</main-class>
+        <main-class>nl.knaw.dans.easy.bagstore.command.Command</main-class>
     </properties>
     <scm>
         <developerConnection>scm:git:https://github.com/DANS-KNAW/easy-bag-store</developerConnection>
@@ -173,6 +173,35 @@
         </plugins>
     </build>
     <profiles>
+        <profile>
+            <id>command</id>
+            <activation>
+                <activeByDefault>true</activeByDefault>
+            </activation>
+            <build>
+                <plugins>
+                    <plugin>
+                        <groupId>org.codehaus.mojo</groupId>
+                        <artifactId>exec-maven-plugin</artifactId>
+                    </plugin>
+                </plugins>
+            </build>
+        </profile>
+        <profile>
+            <id>service</id>
+            <build>
+                <plugins>
+                    <plugin>
+                        <groupId>org.codehaus.mojo</groupId>
+                        <artifactId>exec-maven-plugin</artifactId>
+                        <configuration>
+                            <classpathScope>test</classpathScope>
+                            <mainClass>nl.knaw.dans.easy.bagstore.service.BagStoreService</mainClass>
+                        </configuration>
+                    </plugin>
+                </plugins>
+            </build>
+        </profile>
         <profile>
             <id>rpm</id>
             <activation>
