+++ run-service.sh
@@ -20,8 +20,9 @@ ARGS=$@
 APPHOME=home
 
 MAVEN_OPTS="-agentlib:jdwp=transport=dt_socket,server=y,address=8000,suspend=n" \
-mvn exec:java -Dapp.home=$APPHOME \
-              -Dexec.args="run-service" \
+mvn exec:java -Pservice \
+              -Dapp.home=$APPHOME \
+              -Dexec.args="$ARGS" \
               -Dlogback.configurationFile=$APPHOME/cfg/logback-service.xml \
 #              -Dlogback.statusListenerClass=ch.qos.logback.core.status.OnConsoleStatusListener
