--- src/main/scala/nl.knaw.dans.easy.bagstore/BagStoreVerify.scala
@@ -1,32 +1,0 @@
-/**
- * Copyright (C) 2016 DANS - Data Archiving and Networked Services (info@dans.knaw.nl)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package nl.knaw.dans.easy.bagstore
-
-import scala.util.Try
-
-trait BagStoreVerify { this: BagStoreContext =>
-  type Message = String
-
-  def verify(bagId: Option[BagId] = None): Try[Message] = {
-    // Verify UUID-path components
-    // - Directory name lengths correct at each level
-    // - No stray files
-
-    // Check every bag is virtually-valid
-
-    ???
-  }
-}
