+++ src/test/scala/nl.knaw.dans.easy.bagstore/Bagit4Fixture.scala
@@ -1,0 +1,6 @@
+package nl.knaw.dans.easy.bagstore
+
+trait Bagit4Fixture extends Bagit4FacadeComponent {
+
+  override val bagFacade = new Bagit4Facade()
+}
